implemented crud in user service

+++ src/main/java/com/akash/bloggr/services/UserService.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface UserService {
 
     UserDto createUser(UserDto user);
-    UserDto updateUser(UserDto user, Integer userId);
-    UserDto getUserById(Integer userId);
+    UserDto updateUser(UserDto user, Long userId);
+    UserDto getUserById(Long userId);
     List<UserDto> getAllUsers();
-    void deleteUser(Integer userId);
+    void deleteUser(Long userId);
 
 }
